week2prac/requirements.txt
aws-cdk-lib>=2.219.0
constructs>=10.0.0
